+++ src/App.jsx
@@ -28,39 +28,27 @@ Sin embargo, actualmente Panamá es un país que ya no sólo encuentra su razón
 export default App
 */
 
-import React, { useState } from "react";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
+import './App.css';
+import { HashRouter as Router, Routes, Route } from 'react-router-dom';
+import { Home } from './Pages/Home';
+import { Lugares_a_visitar } from './Pages/Lugares_a_visitar';
 
-const App = () => {
 
-const [theme, setTheme] = useState('light');
-
+function App() {
   return (
-    <div className='container'>
-      <Header theme={theme} setTheme={setTheme}/>
-      <iframe width="560" height="315" src="https://www.youtube.com/embed/jYdaQJzcAcw?si=wHE7c9d9PP9Gown-" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
-      <h1>¡HISTORIA!<br/></h1>
-
-      <p className="read-the-docs">
-      Cruce de caminos, centro de comercio marítimo, unión de dos subcontinentes, Hub de las Américas. Son muchos los eslóganes 
-      que sirven para definir Panamá, y todos aluden a lo mismo: el Canal. Esta maravilla de la ingeniería, aunque 
-      relativamente reciente en su historia, ha puesto al país en el mapa económico mundial. Hace tres millones de años, 
-      tras unos años de movimientos tectónicos y actividad volcánica, surgió una estrecha franja de tierra llamada istmo que separó 
-      las aguas del Caribe y del Pacífico y a la vez unió las masas continentales de América del Norte y del Sur.<br/><br/>
-      </p>
-
-      <p className="read-the-docs">
-      Sin embargo, actualmente Panamá es un país que ya no sólo encuentra su razón de ser en su famoso canal, 
-      si no que se esfuerza por atraer cada vez más turismo. Panamá es mucho más que el canal; es playa, es montaña, 
-      es selva, es cultura, y es, sobre todo, uno de los países con mayor biodiversidad que se hayan visto jamás. 
-      Descubre las maravillas que ofrece explorando este rincón del internet.
-      </p>
-      <Footer/>
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/lugares-a-visitar" element={<Lugares_a_visitar />} />
+        </Routes>
+    </Router>
   )
-}   
+}
+
 export default App
+
+
+
